+++ packages/wasm-bridge/src/types.ts
@@ -1,6 +1,14 @@
 // Copyright 2019-2022 @polkadot/wasm-bridge authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+// Use non-strong types instead of WasmImports which may not
+// be available as part of the TS environment types (it needs dom)
+export type WasmImports = Record<string, (...args: never[]) => unknown>;
+
+// Use non-strong types instead of WebAssembly.Memory which may not
+// be available as part of the TS environment types (it needs dom)
+export type WasmMemory = { buffer: ArrayBuffer };
+
 export declare interface InitResult<C extends WasmBaseInstance> {
   error: string | null;
   type: 'asm' | 'wasm' | 'none';
@@ -9,10 +17,10 @@ export declare interface InitResult<C extends WasmBaseInstance> {
 
 export type InitPromise <C extends WasmBaseInstance> = Promise<InitResult<C>>;
 
-export type InitFn <C extends WasmBaseInstance> = (wbg: WebAssembly.ModuleImports) => InitPromise<C>;
+export type InitFn <C extends WasmBaseInstance> = (wbg: WasmImports) => InitPromise<C>;
 
 export interface BridgeBase<C extends WasmBaseInstance> extends InitResult<C> {
-  wbg: WebAssembly.ModuleImports;
+  wbg: WasmImports;
 
   init (createWasm?: InitFn<C>): Promise<C | null>;
   getObject (idx: number): unknown;
@@ -30,7 +38,7 @@ export interface BridgeBase<C extends WasmBaseInstance> extends InitResult<C> {
 }
 
 export interface WasmBaseInstance {
-  memory: WebAssembly.Memory;
+  memory: WasmMemory;
 
   // wbindgen functions (required and used internally)